add link student tables + advance fixing removing unnused code
